+++ pkg/pillar/pubsub/util.go
@@ -1,0 +1,73 @@
+package pubsub
+
+import (
+	"encoding/json"
+	"errors"
+	"fmt"
+	"io/ioutil"
+	"log"
+	"os"
+	"path/filepath"
+	"reflect"
+	"strings"
+)
+
+func deepCopy(in interface{}) interface{} {
+	b, err := json.Marshal(in)
+	if err != nil {
+		log.Fatal("json Marshal in deepCopy", err)
+	}
+	var output interface{}
+	if err := json.Unmarshal(b, &output); err != nil {
+		log.Fatal(err, "json Unmarshal in deepCopy")
+	}
+	return output
+}
+
+func lookupSlave(slaveCollection localCollection, key string) *interface{} {
+	for slaveKey := range slaveCollection {
+		if slaveKey == key {
+			res := slaveCollection[slaveKey]
+			return &res
+		}
+	}
+	return nil
+}
+
+// TypeToName given a particular object, get the desired name for it
+func TypeToName(something interface{}) string {
+	t := reflect.TypeOf(something)
+	out := strings.Split(t.String(), ".")
+	return out[len(out)-1]
+}
+
+// WriteRename write data to a fmpfile and then rename it to a desired name
+func WriteRename(fileName string, b []byte) error {
+	dirName := filepath.Dir(fileName)
+	// Do atomic rename to avoid partially written files
+	tmpfile, err := ioutil.TempFile(dirName, "pubsub")
+	if err != nil {
+		errStr := fmt.Sprintf("WriteRename(%s): %s",
+			fileName, err)
+		return errors.New(errStr)
+	}
+	defer tmpfile.Close()
+	defer os.Remove(tmpfile.Name())
+	_, err = tmpfile.Write(b)
+	if err != nil {
+		errStr := fmt.Sprintf("WriteRename(%s): %s",
+			fileName, err)
+		return errors.New(errStr)
+	}
+	if err := tmpfile.Close(); err != nil {
+		errStr := fmt.Sprintf("WriteRename(%s): %s",
+			fileName, err)
+		return errors.New(errStr)
+	}
+	if err := os.Rename(tmpfile.Name(), fileName); err != nil {
+		errStr := fmt.Sprintf("WriteRename(%s): %s",
+			fileName, err)
+		return errors.New(errStr)
+	}
+	return nil
+}
